pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "boto3>=1.35.97",
    "django>=5.1.4",
    "django-cleanup>=9.0.0",
    "django-cors-headers>=4.6.0",
    "django-storages>=1.14.4",
    "djangorestframework>=3.15.2",
    "pillow>=11.1.0",
    "psycopg2-binary>=2.9.10",
]
